+++ apps/probability/distribution/chi_squared_distribution.cpp
@@ -33,9 +33,9 @@ float ChiSquaredDistribution::evaluateAtAbscissa(float x) const {
   return std::exp(-lgamma(halfk) - halfX + (halfk-1.0f) * std::log(halfX)) / 2.0f;
 }
 
-bool ChiSquaredDistribution::authorizedValueAtIndex(float x, int index) const {
+bool ChiSquaredDistribution::authorizedValueAtIndex(double x, int index) const {
   assert(index == 0);
-  return x > 0.0f && x == (float)((int)x) && x <= k_maxK;
+  return x > 0.0 && x == (double)((int)x) && x <= k_maxK;
 }
 
 double ChiSquaredDistribution::cumulativeDistributiveFunctionAtAbscissa(double x) const {
